requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# API
fastapi==0.135.1
uvicorn==0.41.0
jinja2==3.1.6
python-multipart==0.0.9

# Auth
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0

# Database
psycopg2-binary==2.9.11

# ML
torch==2.11.0
transformers==5.5.0
accelerate==1.13.0

# AWS
boto3==1.42.84

# Config
pydantic==2.12.5
python-dotenv==1.2.2
